Correct and clean up code

+++ common/algo_lib/include/quick_sort.h
@@ -9,7 +9,7 @@ class QuickSort final : public Sort
 public:
 	std::vector<std::pair<size_t, size_t>> SortList() override;
 private:
-	size_t Partition(int low, int high);
+	int Partition(int low, int high);
 	void QuickSortFunc(int low, int high);
 	std::vector<std::pair<size_t, size_t>> swapList_;
 };
