+++ common/algo_lib/src/quick_sort.cpp
@@ -8,12 +8,12 @@ std::vector<std::pair<size_t, size_t>> QuickSort::SortList()
     return swapList_;
 }
 
-size_t QuickSort::Partition(int low, int high)
+int QuickSort::Partition(int low, int high)
 {
     int pivot = list_[high]; // pivot
-    size_t i = (low - 1); // Index of smaller element and indicates the right position of pivot found so far
+    int i = (low - 1); // Index of smaller element and indicates the right position of pivot found so far
 
-    for (size_t j = low; j <= high - 1; j++)
+    for (int j = low; j <= high - 1; j++)
     {
         // If current element is smaller than the pivot
         if (list_[j] < pivot)
@@ -23,7 +23,7 @@ size_t QuickSort::Partition(int low, int high)
             swapList_.push_back(std::pair<size_t, size_t>(i, j));
         }
     }
-    std::swap(list_[i + 1], list_[high]);
+    std::swap(list_[static_cast<size_t>(i + 1)], list_[high]);
     swapList_.push_back(std::pair<size_t, size_t>(i+1, high));
     return (i + 1);
 }
@@ -34,7 +34,7 @@ void QuickSort::QuickSortFunc(int low, int high)
     {
         /* pi is partitioning index, arr[p] is now
         at right place */
-        size_t pi = Partition(low, high);
+        int pi = Partition(low, high);
 
         // Separately sort elements before
         // partition and after partition
